+++ src/main/java/com/eweware/service/base/payload/GroupTypePayload.java
@@ -36,11 +36,11 @@ public final class GroupTypePayload extends BasePayload implements GroupTypeDAOC
         put(DISPLAY_NAME, displayName);
     }
 
-    public Integer getGroupCount() {
-        return (Integer) get(GROUP_COUNT);
+    public Long getGroupCount() {
+        return (Long) get(GROUP_COUNT);
     }
 
-    public void setGroupCount(Integer groupCount) {
+    public void setGroupCount(Long groupCount) {
         put(GROUP_COUNT, groupCount);
     }
 }
